split test in two files
